backend/pyproject.toml
[project]
name = "margo-backend"
version = "0.1.0"
description = "AI-powered PDF annotation backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.2.0",
    "pymupdf>=1.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
]
